Remove extra_types_for_testing and organise code to match solana-sdk
